requirements.txt
pyexcel_ods3~=0.5.3
flatten-dict==0.3.0
odfpy~=1.4.1
lxml~=4.5.2
ezodf~=0.3.2
requests~=2.24.0